Add class 'ValidationSuccess' and move 'ElementWrapper'
