+++ src/@optimizely/forms-react/src/components/elements/ChoiceElementBlock.tsx
@@ -1,6 +1,6 @@
 import { Choice, equals, isNullOrEmpty } from "@optimizely/forms-sdk";
 import React, { useMemo } from "react";
-import ElementWrapper from "../ElementWrapper";
+import ElementWrapper from "./shared/ElementWrapper";
 import { useElement } from "../../hooks/useElement";
 import { ValidationMessage } from "./shared";
