+++ src/@optimizely/forms-react/src/components/elements/NumberElementBlock.tsx
@@ -1,6 +1,6 @@
 import { Number } from "@optimizely/forms-sdk"
 import React, { useMemo } from "react";
-import ElementWrapper from "../ElementWrapper";
+import ElementWrapper from "./shared/ElementWrapper";
 import { useElement } from "../../hooks/useElement";
 import { ValidationMessage, ElementCaption } from "./shared";
