+++ src/@optimizely/forms-react/src/components/elements/TextareaElementBlock.tsx
@@ -1,6 +1,6 @@
 import { Textarea } from "@optimizely/forms-sdk"
 import React, { useMemo } from "react";
-import ElementWrapper from "../ElementWrapper";
+import ElementWrapper from "./shared/ElementWrapper";
 import { useElement } from "../../hooks/useElement";
 import { ElementCaption, ValidationMessage } from "./shared";
