+++ src/@optimizely/forms-react/src/components/elements/TextboxElementBlock.tsx
@@ -1,6 +1,6 @@
 import { Textbox } from "@optimizely/forms-sdk";
 import React, { useMemo } from "react";
-import ElementWrapper from "../ElementWrapper";
+import ElementWrapper from "./shared/ElementWrapper";
 import { useElement } from "../../hooks/useElement";
 import { ElementCaption, ValidationMessage, DataList } from "./shared";
